legacy/backend/pyproject.toml
[tool.poetry]
name = "pixeltales-backend"
version = "0.1.0"
description = "PixelTales Backend - Where AI Characters Tell Their Stories"
authors = ["Alex"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
langchain = "^0.3.0"
langchain-core = ">=0.3,<0.4"
langchain-community = ">=0.3,<0.4"
langchain-openai = ">=0.2,<0.3"
langchain-anthropic = "^0.3.1"
pydantic = "^2.10.4"
pydantic-settings = "^2.1.0"
redis = "^5.0.1"
python-socketio = "^5.12.1"
sqlalchemy = "^2.0.36"
aiosqlite = "^0.20.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
types-aiofiles = "^24.1.0.20241221"
types-python-dateutil = "^2.9.0.20241206"

[tool.mypy]
plugins = ["pydantic.mypy"]
strict_optional = true
ignore_missing_imports = true
check_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
